requirements.txt
streamlit
pandas
openai
pywin32
python-dotenv
